reorganized files into descriptive directories; added service files to handle jwt and api calls
