+++ react-recipe-receipts/src/components/content/RecipeCard.js
