+++ react-recipe-receipts/src/components/content/RecipeContainer.js
