+++ react-recipe-receipts/src/components/content/TitleCard.js
