+++ react-recipe-receipts/src/components/navigation/Footer.js
